// src/generated-api/src/models/DomainBillingDetails.ts
export interface DomainBillingDetails {
    /**
     * 
     * @type {{ [key: string]: any | undefined; }}
     * @memberof DomainBillingDetails
     */
    additionalDetails?: { [key: string]: any | undefined; };
    /**
     * 
     * @type {DomainBillingAddress}
     * @memberof DomainBillingDetails
     */
    address?: DomainBillingAddress;
    /**
     * Format: "2019-06-01 00:00:00"
     * @type {string}
     * @memberof DomainBillingDetails
     */
    autoInvoiceStartDate?: string;
    /**
     * 
     * @type {boolean}
     * @memberof DomainBillingDetails
     */
    defaultInvoiceIsHidden?: boolean;
    /**
     * 
     * @type {number}
     * @memberof DomainBillingDetails
     */
    defaultPaymentTerms?: number;
    /**
     * 
     * @type {DomainBillingFrequency}
     * @memberof DomainBillingDetails
     */
    frequency?: DomainBillingFrequency;
    /**
     * 
     * @type {number}
     * @memberof DomainBillingDetails
     */
    invoiceAmountThreshold?: number;
    /**
     * 
     * @type {number}
     * @memberof DomainBillingDetails
     */
    invoiceGenerationDaysDelay?: number;
    /**
     * 
     * @type {boolean}
     * @memberof DomainBillingDetails
     */
    isInvoiceCreationAuto?: boolean;
    /**
     * 
     * @type {DomainPaymentDetails}
     * @memberof DomainBillingDetails
     */
    paymentDetails?: DomainPaymentDetails;
    /**
     * 
     * @type {DomainPaymentType}
     * @memberof DomainBillingDetails
     */
    paymentType?: DomainPaymentType;
    /**
     * 
     * @type {DomainBillingSchedule}
     * @memberof DomainBillingDetails
     */
    schedule?: DomainBillingSchedule;
    /**
     * 
     * @type {string}
     * @memberof DomainBillingDetails
     */
    taxId?: string;
}